+++ skyvern/cli/console.py
@@ -1,0 +1,4 @@
+from rich.console import Console
+
+# Global console instance for CLI modules
+console = Console()
